+++ packages/fluent_sdk/lib/src/fluent_registry.dart
@@ -1,45 +1,48 @@
 import 'package:fluent_sdk/fluent_sdk.dart';
 import 'package:get_it/get_it.dart';
 
-class FluentRegistry extends Registry {
+/// Concrete implementation of [Registry] using [GetIt] as the backend engine.
+class FluentRegistry implements Registry {
+  final GetIt _getIt = GetIt.instance;
+
   @override
-  void registerLazySingleton<T extends Object>(
-    T Function(GetIt) factoryFunction, {
+  void allowReassignment({required bool allow}) {
+    _getIt.allowReassignment = allow;
+  }
+
+  @override
+  bool isRegistered<T extends Object>() => _getIt.isRegistered<T>();
+
+  @override
+  void registerFactory<T extends Object>(
+    T Function(GetIt i) factoryFunction, {
     String? instanceName,
   }) {
-    GetIt.instance.registerLazySingleton<T>(
-      () => factoryFunction(GetIt.instance),
+    _getIt.registerFactory<T>(
+      () => factoryFunction(_getIt),
+      instanceName: instanceName,
+    );
+  }
+
+  @override
+  void registerLazySingleton<T extends Object>(
+    T Function(GetIt i) factoryFunction, {
+    String? instanceName,
+  }) {
+    _getIt.registerLazySingleton<T>(
+      () => factoryFunction(_getIt),
       instanceName: instanceName,
     );
   }
 
   @override
   void registerSingleton<T extends Object>(
-    T Function(GetIt) factoryFunction, {
+    T Function(GetIt i) factoryFunction, {
     String? instanceName,
   }) {
-    GetIt.instance.registerSingleton<T>(
-      factoryFunction(GetIt.instance),
+    _getIt.registerSingleton<T>(
+      factoryFunction(_getIt),
       instanceName: instanceName,
     );
   }
-
-  @override
-  void registerFactory<T extends Object>(
-    T Function(GetIt) factoryFunction, {
-    String? instanceName,
-  }) {
-    GetIt.instance.registerFactory(
-      () => factoryFunction(GetIt.instance),
-      instanceName: instanceName,
-    );
-  }
-
-  @override
-  void allowReassignment({required bool allow}) {
-    GetIt.instance.allowReassignment = allow;
-  }
-
-  @override
-  bool isRegistered<T extends Object>() => GetIt.instance.isRegistered<T>();
 }