+++ packages/fluent_localization/fluent_localization/CHANGELOG.md
@@ -58,3 +58,7 @@
 ## 1.2.0
 
 * Upgrade package to flutter version v3.35
+
+## 1.3.0
+
+* Upgrade fluent_sdk version to v0.4.0